+++ CHANGELOG.md
@@ -17,6 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - More logging
 - Fix TUF laptop led power
 - Sanitize the dbus path for aura devices (remove `.` chars)
+- Even more error handling (gracefully)
+- Better checking for dbus interfaces
+- Remove dbus `supported_interfaces`
+- dbus ObjectManager is now at root `/`
 
 ## [v6.0.5]